+++ packages/spl-utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./percent";
-export * from "./metadata";
 export * from "./extendBorsh";
 export * from "./transaction";
 export * from "./token";