+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/ChewingGlassFund/wumbo-programs/compare/v2.2.0...v3.0.0) (2022-01-26)
+
+This was a major version bump because of the upgrade to metaplex mpl-token-metadata and metadata v2. This removed several methods from spl-utils that were token-metadata related.
+
+### Removed
+
+  * Remove the entirety of metadata.ts. Switch to @metaplex-foundation/mpl-token-metadata
+
+### Changed
+
+  * Upgrade to metaplex metadata v2
+
 # [2.2.0](https://github.com/ChewingGlassFund/wumbo-programs/compare/v2.1.11...v2.2.0) (2022-01-25)
 
 ### Added