+++ packages/spl-token-staking/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/StrataFoundation/wumbo/compare/v2.2.0...v3.0.0) (2022-01-26)
+
+**Note:** Version bump only for package @strata-foundation/spl-token-staking
+
+
+
+
+
 # [2.2.0](https://github.com/StrataFoundation/wumbo/compare/v2.1.11...v2.2.0) (2022-01-25)
 
 **Note:** Version bump only for package @strata-foundation/spl-token-staking